Refactor to give listeners semantic interface
This will allow listeners to respond differently in future. An example
given here is that users being removed from an org is shown as green in
the console (for a successful action) but red in HipChat (for a
removal). Another example is allowing action failures to be rendered to
the console but not to HipChat.